added new endpoints updated profile

+++ InclusiveCity/InclusiveCity.Domain/Interfaces/Repositories/IStructureImageRepository.cs
@@ -5,7 +5,7 @@ namespace InclusiveCity.Domain.Interfaces.Repositories
     public interface IStructureImageRepository
     {
         Task AddStructureImage(long osmId, IEnumerable<string> imageUrls);
-        Task<IEnumerable<OsmStructureImage>> GetStructureImages(int osmId);
+        Task<IEnumerable<OsmStructureImage>> GetStructureImages(long osmId);
         Task<Dictionary<long, List<OsmStructureImage>>> GetStructureRangeImages(IEnumerable<long> ids);
     }
 }
